scripts:hotfix - updating install script to the new binary standard (#703)
This pull request contains changes to the new standard binaries names, that were changed in the pr #702.
Now binaries are sepated by archicteture, and this made needed a change to the install script, on the next release the binaries names will change and this pull request contains a validation to the old and the new standard.

Signed-off-by: Nathan Martins <nathan.martins@zup.com.br>
(cherry picked from commit a4bde263fc8ca7b9c77f7ccbfe8ac41461d52323)